+++ Task_1_1_3/src/test/java/ru/nsu/fitkulin/MainTest.java
@@ -2,41 +2,27 @@ package ru.nsu.fitkulin;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
+import java.io.ByteArrayInputStream;
+import java.io.ByteArrayOutputStream;
+import java.io.PrintStream;
+
 import org.junit.jupiter.api.Test;
 
-/**
- * test of all code.
- */
-public class MainTest {
-    /**
-     * first.
-     */
+class MainTest {
     @Test
-    public void testMainMethods() {
-        Expression e = new Add(new Number(4), new Div(new Number(2), new Variable("x")));
+    public void testMain() {
+        ByteArrayInputStream input = new ByteArrayInputStream("(3+7*x-10)\nx = 10\nx".getBytes());
+        System.setIn(input);
+        ByteArrayOutputStream output = new ByteArrayOutputStream();
+        System.setOut(new PrintStream(output));
+        Main.main(new String[0]);
+        String res = output.toString().replaceAll("\\r\\n?", "\n");
 
-        assertEquals("(4+(2/x))", e.toString());
-
-        Expression de = e.derivative("x");
-        assertEquals("(0+(((x*0)-(2*1))/(x*x)))", de.toString());
-
-        double result = e.eval("x = 10; y = 13");
-        assertEquals(4.2, result);
-    }
-
-    /**
-     * second.
-     */
-    @Test
-    public void testMainMethods2() {
-        Expression e = new Sub(new Mul(new Number(2.7), new Variable("x")), new Number(4));
-
-        assertEquals("((2.7*x)-4)", e.toString());
-
-        Expression de = e.derivative("x");
-        assertEquals("(((2.7*1)+(0*x))-0)", de.toString());
-
-        double result = e.eval("x = 10; y = 13");
-        assertEquals(23.0, result);
+        String expectedOutput =
+                "Введите выражение\n" + "Введите переменные в формате 'x = 10; y = 5'\n"
+                        + "Введите переменную, по которой нужно взять производную\n"
+                        + "Выражение: ((3+(7*x))-10)\n"
+                        + "Производная: ((0+((7*1)+(0*x)))-0)\n" + "Результат: 63.0\n";
+        assertEquals(expectedOutput, res);
     }
 }
